# SimpleSocialNetwork/client_api_gen/swagger_api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SimpleSocialNetworkBack",
    "version": "1.0"
  },
  "paths": {
    "/api/Auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new user",
        "requestBody": {
          "description": "User credentials that will be used on login",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialsModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialsModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialsModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login into an account using provided credentials",
        "requestBody": {
          "description": "User credentials",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialsModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialsModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialsModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoggedInUser"
                }
              }
            }
          }
        }
      }
    },
    "/api/Comment": {
      "post": {
        "tags": [
          "Comment"
        ],
        "summary": "Create a new comment in specified post",
        "requestBody": {
          "description": "Comment data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Comment"
        ],
        "summary": "Soft delete comment",
        "parameters": [
          {
            "name": "OpId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "MessageId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Comment/vote": {
      "post": {
        "tags": [
          "Comment"
        ],
        "summary": "Vote on a comment",
        "requestBody": {
          "description": "Vote data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteComment"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteComment"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VoteComment"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/OpMessage": {
      "post": {
        "tags": [
          "OpMessage"
        ],
        "summary": "Create a new post using provided data",
        "requestBody": {
          "description": "Post data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOpMessageModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOpMessageModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOpMessageModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "OpMessage"
        ],
        "summary": "Get a page of posts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page",
            "schema": {
              "type": "integer",
              "description": "Page",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OpMessageModel"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "OpMessage"
        ],
        "summary": "Delete the specifid post",
        "parameters": [
          {
            "name": "postId",
            "in": "query",
            "description": "Id of a post",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Id of a post",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/OpMessage/{id}": {
      "get": {
        "tags": [
          "OpMessage"
        ],
        "summary": "Get a post model with the specified id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of apost",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Id of apost",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpMessageModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/OpMessage/from_user": {
      "get": {
        "tags": [
          "OpMessage"
        ],
        "summary": "Get posts from the specified user",
        "parameters": [
          {
            "name": "login",
            "in": "query",
            "description": "User login to get posts from",
            "required": true,
            "schema": {
              "type": "string",
              "description": "User login to get posts from"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OpMessageModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/OpMessage/comments/{postId}/{page}": {
      "get": {
        "tags": [
          "OpMessage"
        ],
        "summary": "Get comments in a post",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "Post id",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Post id",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "path",
            "description": "Page",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Page",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/OpMessage/exists/{postId}": {
      "get": {
        "tags": [
          "OpMessage"
        ],
        "summary": "Check whether the specified post exists",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "description": "Id of a post",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Id of a post",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/OpMessage/vote": {
      "post": {
        "tags": [
          "OpMessage"
        ],
        "summary": "Vote on a post",
        "requestBody": {
          "description": "Vote data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VotePost"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VotePost"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VotePost"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CredentialsModel": {
        "required": [
          "login",
          "password"
        ],
        "type": "object",
        "properties": {
          "login": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "UserModel": {
        "required": [
          "isAdmin",
          "isDeleted",
          "login"
        ],
        "type": "object",
        "properties": {
          "login": {
            "type": "string"
          },
          "about": {
            "type": "string",
            "nullable": true
          },
          "dateBirth": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "isAdmin": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "LoggedInUser": {
        "required": [
          "login",
          "role",
          "token"
        ],
        "type": "object",
        "properties": {
          "login": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "CreateCommentModel": {
        "required": [
          "content",
          "opId"
        ],
        "type": "object",
        "properties": {
          "opId": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "CommentKeyModel": {
        "required": [
          "messageId",
          "opId"
        ],
        "type": "object",
        "properties": {
          "opId": {
            "type": "integer",
            "format": "int32"
          },
          "messageId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "VoteType": {
        "enum": [
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "VoteComment": {
        "required": [
          "commentId",
          "voteType"
        ],
        "type": "object",
        "properties": {
          "commentId": {
            "$ref": "#/components/schemas/CommentKeyModel"
          },
          "voteType": {
            "$ref": "#/components/schemas/VoteType"
          }
        },
        "additionalProperties": false
      },
      "CreateOpMessageModel": {
        "required": [
          "content",
          "tags",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "OpMessageModel": {
        "required": [
          "content",
          "id",
          "isDeleted",
          "points",
          "sendDate",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "posterId": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "points": {
            "type": "integer",
            "format": "int32"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "sendDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "CommentModel": {
        "required": [
          "content",
          "isDeleted",
          "messageId",
          "opId",
          "points",
          "sendDate"
        ],
        "type": "object",
        "properties": {
          "opId": {
            "type": "integer",
            "format": "int32"
          },
          "messageId": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "string"
          },
          "points": {
            "type": "integer",
            "format": "int32"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "sendDate": {
            "type": "string",
            "format": "date-time"
          },
          "posterId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "VotePost": {
        "required": [
          "postId",
          "voteType"
        ],
        "type": "object",
        "properties": {
          "postId": {
            "type": "integer",
            "format": "int32"
          },
          "voteType": {
            "$ref": "#/components/schemas/VoteType"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}